feat: Consolidate pages and refactor testing structure
 Major refactoring completed:

 Pages Consolidation:
- Merged all individual page files into single pages/pages.ts (1290 lines)
- Removed 11 separate *.page.ts files
- Simplified folder structure from 11 files  1 file

 Code Refactoring:
- Extracted UI text constants to utils/uiText.ts
- Created helper functions in utils/helpers.ts
- Updated baseURL configuration in playwright.config.ts
- Fixed navigation URL issues

 Test Improvements:
- Updated all test imports to use consolidated pages
- Fixed Step 1-7 test compatibility issues
- Enhanced Step 7 results testing with postal code data
- Added backward compatibility methods

 Bug Fixes:
- Fixed age input selector with multiple fallbacks
- Corrected Step6 property names (answerSupport2, answerNursing1-5)
- Fixed Step7 method references (verifyReviewPageLoaded)
- Updated download link verification parameters

 Test Status:
-  Step 1-3, 6-7: Working correctly
-  Step 7 Results: All postal code tests passing
-  Pages consolidation: Complete and functional
-  Some navigation timeouts need optimization

This consolidation significantly improves maintainability while preserving all existing functionality.
